Refactoring completed, mock controllers deleted
