--- CareManagement/Controllers/EmployeesController.cs
@@ -1,169 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using CareManagement.Data;
-using CareManagement.Models.OM;
-
-namespace CareManagement.Controllers
-{
-    public class EmployeesController : Controller
-    {
-        private readonly CareManagementContext _context;
-
-        public EmployeesController(CareManagementContext context)
-        {
-            _context = context;
-        }
-
-        // GET: Employees
-        public async Task<IActionResult> Index()
-        {
-            var careManagementContext = _context.Employee.Include(e => e.Qualification);
-            return View(await careManagementContext.ToListAsync());
-        }
-
-        // GET: Employees/Details/5
-        public async Task<IActionResult> Details(Guid? id)
-        {
-            if (id == null || _context.Employee == null)
-            {
-                return NotFound();
-            }
-
-            var employee = await _context.Employee
-                .Include(e => e.Qualification)
-                .FirstOrDefaultAsync(m => m.EmployeeId == id);
-            if (employee == null)
-            {
-                return NotFound();
-            }
-
-            return View(employee);
-        }
-
-        // GET: Employees/Create
-        public IActionResult Create()
-        {
-            ViewData["QualificationId"] = new SelectList(_context.Qualification, "QualificationId", "QualificationDescription");
-            return View();
-        }
-
-        // POST: Employees/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("EmployeeId,QualificationId,FirstName,LastName,Address,EmergencyContact,Phone,EmployeeType,PayRate,PayType,VacationDays,EmployeeStatus,SickDays,Title,StartDate,TotalHoursWorked")] Employee employee)
-        {
-            if (ModelState.IsValid)
-            {
-                employee.EmployeeId = Guid.NewGuid();
-                _context.Add(employee);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["QualificationId"] = new SelectList(_context.Qualification, "QualificationId", "QualificationDescription", employee.QualificationId);
-            return View(employee);
-        }
-
-        // GET: Employees/Edit/5
-        public async Task<IActionResult> Edit(Guid? id)
-        {
-            if (id == null || _context.Employee == null)
-            {
-                return NotFound();
-            }
-
-            var employee = await _context.Employee.FindAsync(id);
-            if (employee == null)
-            {
-                return NotFound();
-            }
-            ViewData["QualificationId"] = new SelectList(_context.Qualification, "QualificationId", "QualificationDescription", employee.QualificationId);
-            return View(employee);
-        }
-
-        // POST: Employees/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("EmployeeId,QualificationId,FirstName,LastName,Address,EmergencyContact,Phone,EmployeeType,PayRate,PayType,VacationDays,EmployeeStatus,SickDays,Title,StartDate,TotalHoursWorked")] Employee employee)
-        {
-            if (id != employee.EmployeeId)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(employee);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!EmployeeExists(employee.EmployeeId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["QualificationId"] = new SelectList(_context.Qualification, "QualificationId", "QualificationDescription", employee.QualificationId);
-            return View(employee);
-        }
-
-        // GET: Employees/Delete/5
-        public async Task<IActionResult> Delete(Guid? id)
-        {
-            if (id == null || _context.Employee == null)
-            {
-                return NotFound();
-            }
-
-            var employee = await _context.Employee
-                .Include(e => e.Qualification)
-                .FirstOrDefaultAsync(m => m.EmployeeId == id);
-            if (employee == null)
-            {
-                return NotFound();
-            }
-
-            return View(employee);
-        }
-
-        // POST: Employees/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(Guid id)
-        {
-            if (_context.Employee == null)
-            {
-                return Problem("Entity set 'CareManagementContext.Employee'  is null.");
-            }
-            var employee = await _context.Employee.FindAsync(id);
-            if (employee != null)
-            {
-                _context.Employee.Remove(employee);
-            }
-            
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
-
-        private bool EmployeeExists(Guid id)
-        {
-          return (_context.Employee?.Any(e => e.EmployeeId == id)).GetValueOrDefault();
-        }
-    }
-}
