--- CareManagement/Views/Employees/Edit.cshtml
@@ -1,104 +1,0 @@
-@model CareManagement.Models.OM.Employee
-
-@{
-    ViewData["Title"] = "Edit";
-}
-
-<h1>Edit</h1>
-
-<h4>Employee</h4>
-<hr />
-<div class="row">
-    <div class="col-md-4">
-        <form asp-action="Edit">
-            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            <input type="hidden" asp-for="EmployeeId" />
-            <div class="form-group">
-                <label asp-for="QualificationId" class="control-label"></label>
-                <select asp-for="QualificationId" class="form-control" asp-items="ViewBag.QualificationId"></select>
-                <span asp-validation-for="QualificationId" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="FirstName" class="control-label"></label>
-                <input asp-for="FirstName" class="form-control" />
-                <span asp-validation-for="FirstName" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="LastName" class="control-label"></label>
-                <input asp-for="LastName" class="form-control" />
-                <span asp-validation-for="LastName" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Address" class="control-label"></label>
-                <input asp-for="Address" class="form-control" />
-                <span asp-validation-for="Address" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="EmergencyContact" class="control-label"></label>
-                <input asp-for="EmergencyContact" class="form-control" />
-                <span asp-validation-for="EmergencyContact" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Phone" class="control-label"></label>
-                <input asp-for="Phone" class="form-control" />
-                <span asp-validation-for="Phone" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="EmployeeType" class="control-label"></label>
-                <select asp-for="EmployeeType" class="form-control"></select>
-                <span asp-validation-for="EmployeeType" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="PayRate" class="control-label"></label>
-                <input asp-for="PayRate" class="form-control" />
-                <span asp-validation-for="PayRate" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="PayType" class="control-label"></label>
-                <select asp-for="PayType" class="form-control"></select>
-                <span asp-validation-for="PayType" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="VacationDays" class="control-label"></label>
-                <input asp-for="VacationDays" class="form-control" />
-                <span asp-validation-for="VacationDays" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="EmployeeStatus" class="control-label"></label>
-                <select asp-for="EmployeeStatus" class="form-control"></select>
-                <span asp-validation-for="EmployeeStatus" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="SickDays" class="control-label"></label>
-                <input asp-for="SickDays" class="form-control" />
-                <span asp-validation-for="SickDays" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Title" class="control-label"></label>
-                <select asp-for="Title" class="form-control"></select>
-                <span asp-validation-for="Title" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="StartDate" class="control-label"></label>
-                <input asp-for="StartDate" class="form-control" />
-                <span asp-validation-for="StartDate" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="TotalHoursWorked" class="control-label"></label>
-                <input asp-for="TotalHoursWorked" class="form-control" />
-                <span asp-validation-for="TotalHoursWorked" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <input type="submit" value="Save" class="btn btn-primary" />
-            </div>
-        </form>
-    </div>
-</div>
-
-<div>
-    <a asp-action="Index">Back to List</a>
-</div>
-
-@section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
-}
