--- CareManagement/Views/Employees/Index.cshtml
@@ -1,119 +1,0 @@
-@model IEnumerable<CareManagement.Models.OM.Employee>
-
-@{
-    ViewData["Title"] = "Index";
-}
-
-<h1>Index</h1>
-
-<p>
-    <a asp-action="Create">Create New</a>
-</p>
-<table class="table">
-    <thead>
-        <tr>
-            <th>
-                @Html.DisplayNameFor(model => model.Qualification)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.FirstName)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.LastName)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Address)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.EmergencyContact)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Phone)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.EmployeeType)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.PayRate)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.PayType)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.VacationDays)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.EmployeeStatus)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.SickDays)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Title)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.StartDate)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.TotalHoursWorked)
-            </th>
-            <th></th>
-        </tr>
-    </thead>
-    <tbody>
-@foreach (var item in Model) {
-        <tr>
-            <td>
-                @Html.DisplayFor(modelItem => item.Qualification.QualificationDescription)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.FirstName)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.LastName)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Address)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.EmergencyContact)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Phone)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.EmployeeType)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.PayRate)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.PayType)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.VacationDays)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.EmployeeStatus)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.SickDays)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Title)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.StartDate)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.TotalHoursWorked)
-            </td>
-            <td>
-                <a asp-action="Edit" asp-route-id="@item.EmployeeId">Edit</a> |
-                <a asp-action="Details" asp-route-id="@item.EmployeeId">Details</a> |
-                <a asp-action="Delete" asp-route-id="@item.EmployeeId">Delete</a>
-            </td>
-        </tr>
-}
-    </tbody>
-</table>
