+++ CareManagement/Views/Schedules/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<CareManagement.Controllers.ScheduleViewModel>
+@model IEnumerable<CareManagement.Controllers.SCHDL.ScheduleViewModel>
 
 @{
     ViewData["Title"] = "Index";
