+++ CareManagement/Views/Shared/_Layout.cshtml
@@ -33,15 +33,6 @@
                             <a class="nav-link text-dark" asp-area="" asp-controller="Qualifications" asp-action="Index">Qualifications</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Employees" asp-action="Index">Employees</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Shifts" asp-action="Index">Shifts</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Renters" asp-action="Index">Renters</a>
-                        </li>
-                        <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="Schedules" asp-action="Index">Schedules</a>
                         </li>
                         <li class="nav-item">
