--- CareManagement/Views/Shifts/Index.cshtml
@@ -1,65 +1,0 @@
-@model IEnumerable<CareManagement.Models.OM.Shift>
-
-@{
-    ViewData["Title"] = "Index";
-}
-
-<h1>Index</h1>
-
-<p>
-    <a asp-action="Create">Create New</a>
-</p>
-<table class="table">
-    <thead>
-        <tr>
-            <th>
-                @Html.DisplayNameFor(model => model.EmployeeId)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.ManagerId)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.StartTime)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.EndTime)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Sick)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.ScheduleId)
-            </th>
-            <th></th>
-        </tr>
-    </thead>
-    <tbody>
-@foreach (var item in Model) {
-        <tr>
-            <td>
-                @Html.DisplayFor(modelItem => item.EmployeeId)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.ManagerId)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.StartTime)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.EndTime)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Sick)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.ScheduleId)
-            </td>
-            <td>
-                <a asp-action="Edit" asp-route-id="@item.ShiftId">Edit</a> |
-                <a asp-action="Details" asp-route-id="@item.ShiftId">Details</a> |
-                <a asp-action="Delete" asp-route-id="@item.ShiftId">Delete</a>
-            </td>
-        </tr>
-}
-    </tbody>
-</table>
